+++ packages/react-sprucebot/lib/components/ButtonGroup/ButtonGroup.js
@@ -1,7 +1,5 @@
 "use strict";
 
-var _interopRequireWildcard = require("@babel/runtime/helpers/interopRequireWildcard");
-
 var _interopRequireDefault = require("@babel/runtime/helpers/interopRequireDefault");
 
 Object.defineProperty(exports, "__esModule", {
@@ -15,7 +13,7 @@ var _react = _interopRequireDefault(require("react"));
 
 var _classnames = _interopRequireDefault(require("classnames"));
 
-var _Button = _interopRequireWildcard(require("../Button/Button"));
+var _Button = _interopRequireDefault(require("../Button/Button"));
 
 var ButtonGroup = function ButtonGroup(props) {
   var actions = props.actions,
@@ -38,7 +36,9 @@ var ButtonGroup = function ButtonGroup(props) {
     }, _react.default.createElement(_Button.default, (0, _extends2.default)({
       kind: kind ? btnKindKey[kind] : '',
       isFullWidth: kind === 'floating'
-    }, action)));
+    }, action, {
+      kind: kind === 'floating' ? 'simple' : kind === 'segmented' ? 'secondary' : action.kind
+    })));
   }));
 };