+++ packages/react-sprucebot/lib/components/Tabs/components/Tab/Tab.js
@@ -20,9 +20,10 @@ var _Button = _interopRequireDefault(require("../../../Button/Button"));
 var Tab = function Tab(props) {
   var text = props.text,
       isCurrent = props.isCurrent,
-      rest = (0, _objectWithoutProperties2.default)(props, ["text", "isCurrent"]);
+      className = props.className,
+      rest = (0, _objectWithoutProperties2.default)(props, ["text", "isCurrent", "className"]);
   return _react.default.createElement("li", {
-    className: "tab"
+    className: (0, _classnames.default)('tab', className)
   }, _react.default.createElement(_Button.default, (0, _extends2.default)({
     className: (0, _classnames.default)('tab__inner', {
       'tab--is-current': props.isCurrent