+++ packages/react-sprucebot/lib/components/Toast/components/ToastWrapper/ToastWrapper.js
@@ -15,14 +15,11 @@ var _react = _interopRequireWildcard(require("react"));
 
 var _velocityReact = require("velocity-react");
 
-var _Toast = _interopRequireWildcard(require("../../Toast"));
+var _Toast = _interopRequireDefault(require("../../Toast"));
 
 var ToastWrapper = function ToastWrapper(props) {
   var toasts = props.toasts,
       handleRemove = props.handleRemove;
-  console.log({
-    toasts: toasts
-  });
   return _react.default.createElement("div", {
     className: "toasts-wrapper"
   }, _react.default.createElement(_velocityReact.VelocityTransitionGroup, {