+++ packages/sprucebot-skills-kit/config/default.js
@@ -3,6 +3,7 @@ const path = require('path')
 const { omit, pick } = require('lodash')
 const fs = require('fs')
 const errors = require('./errors')
+const packageJSON = require('../package.json')
 // Check for .env
 try {
 	require('dotenv').config()
@@ -12,6 +13,19 @@ try {
 
 module.exports = {
 	DEV_MODE: process.env.DEV_MODE === 'true',
+	ENV: process.env.ENV || 'default',
+	PACKAGE_NAME: packageJSON.name,
+	PACKAGE_VERSION: packageJSON.version,
+	LOG_LEVEL: process.env.LOG_LEVEL || 'warn',
+	METRICS_APP_KEY: process.env.METRICS_APP_KEY,
+	METRICS_URL: process.env.METRICS_URL,
+	METRICS_ENABLED: process.env.METRICS_ENABLED === 'true',
+	METRICS_BROWSER_STATS_ENABLED:
+		process.env.METRICS_BROWSER_STATS_ENABLED === 'true',
+	METRICS_REQUESTS_DISABLED: process.env.METRICS_REQUESTS_DISABLED === 'true',
+	METRICS_SERVER_STATS_DISABLED:
+		process.env.METRICS_SERVER_STATS_DISABLED === 'true',
+	METRICS_SEQUELIZE_DISABLED: process.env.METRICS_SEQUELIZE_DISABLED === 'true',
 	API_HOST: process.env.API_HOST,
 	API_KEY: process.env.API_KEY,
 	SKILL_STYLESHEET: process.env.SKILL_STYLESHEET,
@@ -28,9 +42,7 @@ module.exports = {
 		process.env.INTERFACE_SSL_ALLOW_SELF_SIGNED === 'true',
 	API_SSL_ALLOW_SELF_SIGNED: process.env.API_SSL_ALLOW_SELF_SIGNED === 'true',
 	WHITELABEL: process.env.WHITELABEL,
-	REDIS_URL: process.env.REDIS_URL || null,
-	REDIS_DEFAULT_TTL_SEC: process.env.REDIS_DEFAULT_TTL_SEC || 300,
-	REDIS_DISABLE: process.env.REDIS_DISABLE === 'true',
+
 	RUN_CRONS: process.env.RUN_CRONS === 'true',
 	ENABLE_DEBUG_ROUTES: process.env.ENABLE_DEBUG_ROUTES === 'true',
 	// Event contract
@@ -104,6 +116,14 @@ module.exports = {
 				accessKeyId: process.env.AWS_ACCESS_KEY_ID,
 				secretAccessKey: process.env.AWS_SECRET_ACCESS_KEY
 			}
+		},
+		cache: {
+			cache: './cache/redis',
+			enable: process.env.CACHE_ENABLE === 'true',
+			options: {
+				url: process.env.REDIS_URL || null,
+				ttl: process.env.DEFAULT_TTL_SEC || 300
+			}
 		}
 	}, // Settings for any services.
 	bodyParserOptions: {
@@ -139,6 +159,13 @@ module.exports = {
 			'log_colors',
 			'nextConfig',
 			'WHITELABEL',
-			'SLUG'
+			'SLUG',
+			'PACKAGE_NAME',
+			'PACKAGE_VERSION',
+			'LOG_LEVEL',
+			'ENV',
+			'METRICS_URL',
+			'METRICS_ENABLED',
+			'METRICS_BROWSER_STATS_ENABLED'
 		])
 }