+++ packages/sprucebot-skills-kit/server/controllers/1.0/teammate/index.js
@@ -17,6 +17,7 @@ module.exports = router => {
 			}
 
 			const users = await ctx.sb.users(ctx.auth.Location.id, params)
+
 			ctx.body = users
 			await next()
 		}