+++ CHANGELOG.md
@@ -3,13 +3,82 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-## [7.4.1](https://github.com/sprucelabsai/workspace.sprucebot-skills-kit/compare/v7.4.0...v7.4.1) (2018-10-03)
+## [7.7.3](https://github.com/sprucelabsai/workspace.sprucebot-skills-kit/compare/v7.7.2...v7.7.3) (2018-10-31)
 
 
 ### Bug Fixes
 
-* move swagger-jsdoc to dependency ([170eca0](https://github.com/sprucelabsai/workspace.sprucebot-skills-kit/commit/170eca0))
-* Remove 404ing next stylesheet ([ab6e0fd](https://github.com/sprucelabsai/workspace.sprucebot-skills-kit/commit/ab6e0fd))
+* swagger doc should generate after sprucebot-skills-kit-server initalized ([38d097c](https://github.com/sprucelabsai/workspace.sprucebot-skills-kit/commit/38d097c))
+
+
+
+
+
+## [7.7.2](https://github.com/sprucelabsai/workspace.sprucebot-skills-kit/compare/v7.7.1...v7.7.2) (2018-10-24)
+
+
+### Testing
+
+* mock global logger ([a9dbb15](https://github.com/sprucelabsai/workspace.sprucebot-skills-kit/commit/a9dbb15))
+
+
+
+
+
+## [7.7.1](https://github.com/sprucelabsai/workspace.sprucebot-skills-kit/compare/v7.7.0...v7.7.1) (2018-10-23)
+
+**Note:** Version bump only for package workspace.sprucebot-skills-kit
+
+
+
+
+
+# [7.7.0](https://github.com/sprucelabsai/workspace.sprucebot-skills-kit/compare/v7.6.1...v7.7.0) (2018-10-23)
+
+
+### Features
+
+* browser metrics; make ctx.sb.audit handle its own error and no await ([c9b0cf0](https://github.com/sprucelabsai/workspace.sprucebot-skills-kit/commit/c9b0cf0))
+
+
+
+
+
+## [7.6.1](https://github.com/sprucelabsai/workspace.sprucebot-skills-kit/compare/v7.6.0...v7.6.1) (2018-10-22)
+
+
+### Bug Fixes
+
+* remove duplicate in config ([45e5ccc](https://github.com/sprucelabsai/workspace.sprucebot-skills-kit/commit/45e5ccc))
+
+
+
+
+
+# [7.6.0](https://github.com/sprucelabsai/workspace.sprucebot-skills-kit/compare/v7.5.0...v7.6.0) (2018-10-21)
+
+
+### Documentation
+
+* Add audit logging doc with example usage ([fd3f607](https://github.com/sprucelabsai/workspace.sprucebot-skills-kit/commit/fd3f607))
+* update .env example ([e1c9726](https://github.com/sprucelabsai/workspace.sprucebot-skills-kit/commit/e1c9726))
+
+
+### Features
+
+* Implement sprucebot logger and audit log ([3db6d36](https://github.com/sprucelabsai/workspace.sprucebot-skills-kit/commit/3db6d36))
+* metrics log info for enabled/disabled features ([b139f7a](https://github.com/sprucelabsai/workspace.sprucebot-skills-kit/commit/b139f7a))
+
+
+
+
+
+# [7.5.0](https://github.com/sprucelabsai/workspace.sprucebot-skills-kit/compare/v7.4.4...v7.5.0) (2018-10-08)
+
+
+### Features
+
+* standardized caching ([49e5137](https://github.com/sprucelabsai/workspace.sprucebot-skills-kit/commit/49e5137))